feat: make-quiz category protector 적용
