+++ src/components/left-sidebar.tsx
@@ -126,7 +126,7 @@ function LogoIcon() {
         d="M30.4362 1.57121C30.4362 1.57121 30.4014 1.57681 30.3857 1.58356C23.177 3.41289 15.4999 5.96051 9.82236 10.9984C7.41027 13.1428 5.49169 15.9363 5.29247 19.2525C4.98085 24.4396 8.91567 28.8968 14.0803 29.2071C16.2065 29.3348 18.339 28.7203 20.0763 27.478C21.7457 26.2817 22.883 24.5032 23.8096 22.662C25.2415 19.8084 29.2221 12.0253 29.2221 12.0253C29.3606 11.7604 29.0989 11.46 28.8203 11.5702C27.5011 11.8488 24.5427 12.7561 23.411 13.4192C23.2561 13.5099 23.0698 13.3602 23.1221 13.1902C23.6245 11.5236 24.7717 9.89956 25.6757 8.49947C27.184 6.1584 28.8831 4.02117 30.8005 2.01248C30.8617 1.95075 30.9235 1.87751 30.9288 1.78933C30.9415 1.57848 30.6528 1.53805 30.436 1.57504L30.4362 1.57121Z"
         fill="url(#paint0_radial_3978_53)"
       />
-      <g clip-path="url(#clip0_3978_53)">
+      <g clipPath="url(#clip0_3978_53)">
         <path
           d="M50.5441 10.8444C49.6078 10.2343 48.47 9.92383 47.1471 9.92383C46.2489 9.92383 45.4541 10.0491 44.7573 10.3052C44.0605 10.5612 43.4453 10.8935 42.9064 11.302C42.3674 11.7106 41.8829 12.1682 41.4529 12.6802L41.6053 10.2888H38.7908L34.8984 32.4763H37.9579L39.5802 23.3464C39.8251 23.8584 40.179 24.316 40.6363 24.7246C41.0936 25.1332 41.6488 25.4654 42.3021 25.7215C42.9554 25.9775 43.6685 26.1028 44.447 26.1028C45.6718 26.1028 46.8042 25.874 47.844 25.411C48.8837 24.9534 49.7983 24.3106 50.5822 23.4826C51.3661 22.6546 51.9976 21.6849 52.4658 20.5736C52.934 19.4623 53.1898 18.2476 53.2334 16.9402C53.315 15.611 53.1191 14.418 52.6509 13.3557C52.1827 12.2935 51.475 11.4546 50.5387 10.8444H50.5441ZM49.2594 20.481C48.7476 21.4125 48.078 22.137 47.256 22.6546C46.434 23.1721 45.4976 23.4335 44.4524 23.4335C43.6141 23.4335 42.8846 23.2374 42.264 22.8507C41.6434 22.4639 41.1698 21.9246 40.854 21.2437C40.5383 20.5573 40.4076 19.7456 40.473 18.8086C40.5111 17.6048 40.7887 16.5425 41.3004 15.6219C41.8121 14.7013 42.4872 13.9658 43.3364 13.4157C44.1802 12.8655 45.1275 12.5876 46.1672 12.5876C47.0219 12.5876 47.7514 12.7946 48.3393 13.1978C48.9327 13.6063 49.3846 14.162 49.7003 14.8701C50.0161 15.5729 50.1521 16.3845 50.114 17.3051C50.0542 18.4872 49.7656 19.5495 49.2594 20.4756V20.481Z"
           fill="#292B2C"
@@ -170,8 +170,8 @@ function LogoIcon() {
           gradientUnits="userSpaceOnUse"
           gradientTransform="translate(8.21943 26.6324) rotate(-68.4287) scale(16.6317 14.2798)"
         >
-          <stop stop-color="#C7D6FF" />
-          <stop offset="1" stop-color="#FB7E20" />
+          <stop stopColor="#C7D6FF" />
+          <stop offset="1" stopColor="#FB7E20" />
         </radialGradient>
         <clipPath id="clip0_3978_53">
           <rect
